Use asynchronous read of data pipe from asciidoctor process
Previously, data was not read from the pipe until after the asciidoctor
process finished. The data pipe has a buffer of size 64KB and once full,
blocks execution until data is read. If the HTML file corresponding to
the AsciiDoc file exceeded the buffer size, then the end result was that
the Quick Look preview was never generated. Now the code reads the data
asynchronously from the pipe.

Fixes #1